Restructure HTTP,WS library functions
There are three major changes in this commit:

Firstly, we changed the HTTP functions so they start a new session (and
so a new stream) for each request. This makes implementation easier, but
will scale worse, so we should look into adding keepalive functionality
(from HTTP 1.1 I believe) somewhere down the line.

Secondly, we made the websocket library features build upon the http
stream functionality, rather than directly using beast for each step.
This is more in-line with how websocket's overall protocol, so it should
increase ease-of-use and maintainability.

Thirdly, we templatized a lot of the HTTP/WS functions. This has the
downside of requiring additional `_impl.hpp` headers (commonly labeled
.tpp in things like compiler implementations), but hopefully gets us
started on a somewhat useful level of genericity. As of now, though,
this is not the case, as many functions require explicit instantiation
for template deduction to succeed...