Redocumentation, add new gateway class
This change renames the previous discpp::connection class to be under
the discpp::gateway namespace, which establishes its role as a gateway
connection object. As such, construction is used as a mechanism to
connect to the discord gateway endpoint, and its members currently
handle automatically reading and keeping the connection alive.

Later we might want to add in user-facing options for whether such
keep-alive behavior should be automatic, but there are a lot of
functions that should be user-provided/requested, so we still have time
to decide on these all at once.

Another (less apparent) feature is the partial addition of Doxygen style
documentation. This should help users of the library (as well as
maintainers) keep track of features and interfaces much easier, as well
as help us collect known bugs and issues more closely with the code.