Remove out-of-scope abstractions
The scope of this library should be to just provide a shimming layer
between the Discord HTTP/WebSocket APIs and a C++ program that properly
manages the state of the connection and any higher level actions on the
connection.

This seems to be the solution to the complicated heartbeating problem we
had before. All messages in and out should look the same to the
connection interface, and the logic for heartbeating should occur above
this layer. Special thanks to friends suggesting more abstraction! :)