Switch to priority queue based messaging.
The priority queue is implemented using a deadline-based priority
system; each message object is actually a pair composed of 1.) the
actual JSON gateway object, and 2.) a boost::optional-wrapped
time_point using the standard library's steady_clock. This way, objects
with closer deadlines are guaranteed to be passed first.

The rationale for using steady_clock is as follows:
    - none of the connection-critical commands (e.g. heartbeating)
      require the notion of absolute time
    - we *need* monotonicity to prevent time travel issues
    - we do not presently have a need for high precision time tracking

In the future we should probably templatize the class to allow things
like custom choices for Clock, Compare, etc options with sane defaults
(that the user won't need to change 99% of the time).