Migrate from nlohmann::json to boost::json
Our mentality for migrating comes from mostly a speed vs complexity
concern. nlohmann::json values ease-of-use over memory usage and speed,
which isn't particularly something we agree with; boost::json aims for a
level of optimization we would prefer, at the cost of losing the level
of this "It Just Works" behavior.

Naturally, we also would like the ability to customize our own usage of
the data type more, and the incremental parsing and serialization of
boost::json is pretty good about that, too.