Expose some members and add deadline tagging
This should make the core message processing functional yet again, and
this time properly (for the most part). In retrospect, our priority
assignments aren't quite perfect (since establishing a specific
time-based deadline isn't as important as relative priority, and thus
could cause a performance or memory efficiency overhead).

We expose the underlying message queues to allow library users to
directly utilize the queue waiting functionality; this might not be a
sensible choice; it could be beneficial to only expose waiting
functionality, or rethink the organization altogether.